quantum_engine.py
#!/usr/bin/env python3
"""
Quantum Chess AI Engine
Simulates quantum-inspired decision making for chess moves.
"""

import json
import sys
import numpy as np
import chess
import random
from typing import Dict, List, Tuple, Optional

class QuantumChessEngine:
    def __init__(self):
        self.shots = 1000
        
    def encode_board_state(self, fen: str) -> np.ndarray:
        """Encode chess board state into a quantum-inspired representation."""
        board = chess.Board(fen)
        state_vector = np.zeros(64)  # 64 squares
        
        for square in chess.SQUARES:
            piece = board.piece_at(square)
            if piece:
                # Encode piece type and color
                piece_value = self._get_piece_value(piece)
                color_factor = 1 if piece.color else -1
                state_vector[square] = piece_value * color_factor
                
        return state_vector
    
    def _get_piece_value(self, piece) -> float:
        """Get quantum-encoded value for a chess piece."""
        piece_values = {
            chess.PAWN: 1.0,
            chess.KNIGHT: 3.0,
            chess.BISHOP: 3.2,
            chess.ROOK: 5.0,
            chess.QUEEN: 9.0,
            chess.KING: 0.0  # King has special quantum properties
        }
        return piece_values.get(piece.piece_type, 0.0)
    
    def create_quantum_simulation(self, board_state: np.ndarray) -> np.ndarray:
        """Create a quantum-inspired simulation of the chess position."""
        # Simulate quantum superposition using probability distributions
        normalized_state = board_state / (np.linalg.norm(board_state) + 1e-10)
        
        # Apply quantum-inspired transformations
        # Simulate quantum Fourier transform effect
        fft_state = np.fft.fft(normalized_state)
        
        # Add quantum noise/uncertainty
        noise = np.random.normal(0, 0.1, len(fft_state))
        quantum_state = fft_state + noise
        
        # Normalize to create probability distribution
        probabilities = np.abs(quantum_state) ** 2
        probabilities = probabilities / np.sum(probabilities)
        
        return probabilities
    
    def quantum_move_evaluation(self, fen: str, move: str) -> float:
        """Evaluate a move using quantum-inspired algorithms."""
        board = chess.Board(fen)
        
        # Make the move
        move_obj = chess.Move.from_uci(move)
        board.push(move_obj)
        
        # Encode the resulting position
        board_state = self.encode_board_state(board.fen())
        
        # Create quantum simulation
        probabilities = self.create_quantum_simulation(board_state)
        
        # Calculate quantum evaluation
        quantum_score = 0.0
        
        for square in range(64):
            # Weight by piece value at that square
            piece_value = abs(board_state[square])
            probability = probabilities[square]
            quantum_score += piece_value * probability
        
        return quantum_score
    
    def generate_quantum_move(self, fen: str) -> Dict:
        """Generate a quantum-inspired move suggestion."""
        board = chess.Board(fen)
        legal_moves = list(board.legal_moves)
        
        if not legal_moves:
            return {
                "move": "",
                "concept": "superposition",
                "probability": 0.0,
                "reason": "No legal moves available"
            }
        
        # Evaluate all moves using quantum algorithm
        move_evaluations = []
        for move in legal_moves:
            move_str = move.uci()
            evaluation = self.quantum_move_evaluation(fen, move_str)
            move_evaluations.append((move_str, evaluation))
        
        # Convert evaluations to probabilities
        evaluations = np.array([eval for _, eval in move_evaluations])
        # Apply softmax to get probabilities
        exp_evaluations = np.exp(evaluations - np.max(evaluations))
        probabilities = exp_evaluations / np.sum(exp_evaluations)
        
        # Select move based on quantum probability distribution
        selected_idx = np.random.choice(len(legal_moves), p=probabilities)
        selected_move, selected_eval = move_evaluations[selected_idx]
        selected_prob = probabilities[selected_idx]
        
        # Determine quantum concept
        concept = self._determine_quantum_concept(board, selected_move)
        
        return {
            "move": selected_move,
            "concept": concept,
            "probability": float(selected_prob),
            "evaluation": float(selected_eval)
        }
    
    def _determine_quantum_concept(self, board: chess.Board, move: str) -> str:
        """Determine which quantum concept best describes the move."""
        move_obj = chess.Move.from_uci(move)
        
        # Check if it's a capture
        if board.is_capture(move_obj):
            return "probabilistic_collapse"
        
        # Check if it's a check
        board_copy = board.copy()
        board_copy.push(move_obj)
        if board_copy.is_check():
            return "quantum_tunneling"
        
        # Check if it's a center move
        center_squares = [chess.D4, chess.D5, chess.E4, chess.E5]
        if move_obj.to_square in center_squares:
            return "coherence"
        
        # Check if it's a queen or rook move
        piece = board.piece_at(move_obj.from_square)
        if piece and piece.piece_type in [chess.QUEEN, chess.ROOK]:
            return "entanglement"
        
        return "superposition"
    
    def get_quantum_explanation(self, concept: str) -> str:
        """Get explanation for quantum concept."""
        explanations = {
            "superposition": "Maintains multiple tactical pathways simultaneously",
            "entanglement": "Coordinates piece threats in tandem across the board",
            "probabilistic_collapse": "Calculated risk that can yield sudden advantage",
            "quantum_tunneling": "Bypasses traditional defensive structures",
            "coherence": "Maintains strategic harmony across all pieces"
        }
        return explanations.get(concept, "Quantum-inspired strategic thinking")

def main():
    """Main function to handle command line interface."""
    if len(sys.argv) != 2:
        print("Usage: python quantum_engine.py <fen_string>")
        sys.exit(1)
    
    fen = sys.argv[1]
    engine = QuantumChessEngine()
    
    try:
        result = engine.generate_quantum_move(fen)
        print(json.dumps(result))
    except Exception as e:
        print(json.dumps({
            "error": str(e),
            "move": "",
            "concept": "superposition",
            "probability": 0.0
        }))

if __name__ == "__main__":
    main()
