requirements.txt
qiskit>=0.40.0
numpy>=1.21.0
python-chess>=1.9.0
flask>=2.0.0
flask-cors>=3.0.0
